feat(cli): create graphql schema using gql.tada
