+++ .changeset/rotten-cheetahs-bathe.md
@@ -1,0 +1,6 @@
+---
+"@bigcommerce/create-catalyst": minor
+"@bigcommerce/catalyst-core": minor
+---
+
+create graphql schema using gql.tada
