+++ apps/core/package.json
@@ -4,11 +4,12 @@
   "version": "0.1.1",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"npm:dev-*\"",
+    "dev": "npm run generate && concurrently \"npm:dev-*\"",
     "dev-next": "next dev",
     "dev-codegen": "dotenv -e .env.local -- npm run codegen -- -w",
     "codegen": "dotenv -e .env.local -- gql-gen --errors-only --config codegen.ts",
-    "build": "npm run codegen && next build",
+    "generate": "dotenv -e .env.local -- node ./scripts/generate.cjs",
+    "build": "npm run generate && npm run codegen && next build",
     "start": "next start",
     "lint": "next lint",
     "typecheck": "tsc --noEmit",
@@ -23,7 +24,7 @@
     "@vercel/kv": "^1.0.0",
     "@vercel/speed-insights": "^1.0.1",
     "clsx": "^2.0.0",
-    "gql.tada": "^1.3.2",
+    "gql.tada": "^1.3.5",
     "graphql": "^16.8.1",
     "lodash.debounce": "^4.0.8",
     "lucide-react": "^0.294.0",
@@ -44,6 +45,7 @@
   "devDependencies": {
     "@0no-co/graphqlsp": "^1.5.0",
     "@bigcommerce/eslint-config-catalyst": "workspace:^",
+    "@gql.tada/cli-utils": "^0.2.0",
     "@graphql-codegen/cli": "^5.0.0",
     "@graphql-codegen/client-preset": "^4.1.0",
     "@graphql-codegen/schema-ast": "^4.0.0",
